refactor(compare-file-size): put calculating file sizes in own module
