+++ lib/compare-file-size.js
@@ -1,9 +1,8 @@
 'use strict';
 
-const fs = require('fs');
-const path = require('path');
 const promisify = require('bluebird').promisify;
 const glob = promisify(require('glob').glob);
+const getFileSizesPerType = require('./get-filesizes-per-type');
 
 function compareFileSize(sourceFolder, tempFolder) {
 	return glob('**/*', { nodir: true, cwd: sourceFolder })
@@ -15,28 +14,6 @@ function compareFileSize(sourceFolder, tempFolder) {
 				dest: getFileSizesPerType(tempFiles)
 			}
 		})
-}
-
-function getFileSizesPerType(files) {
-	const sizes = getFileSizes(files);
-	return Object.keys(sizes).reduce((typeSizes, file) => {
-		const fileType = path.extname(file).replace('.', '');
-		const fileSize = sizes[file];
-
-		if (!typeSizes.hasOwnProperty(fileType)) {
-			typeSizes[fileType] = 0;
-		}
-
-		typeSizes[fileType] += fileSize;
-		return typeSizes;
-	}, {});
-}
-
-function getFileSizes(files) {
-	return files.reduce((sizes, file) => {
-		sizes[file] = fs.statSync(file).size;
-		return sizes;
-	}, {});
 }
 
 function appendDirToFiles(dir, files) {
